Replace backend_manager, adjust tests
Replace the usage of the functions in `_backend_manager` with:
* for `QuantumProgram`, the equivalent functions from `qiskit.wizard`,
  with the assumption that users via the to-be-deprecated
  `QuantumProgram` would accept that only one connection to the API
  (or none) is allowed. The same assumption is used in the tests.
* for the rest of the SDK, use the `Backend` objects directly. The
  same principle is applied to the tests.

For decoupling, adjust the signature of several functions and methods
in order to accept `Backend` objects directly, instead of receiving
the name. Note that this is also applied to
`QJob.qobj['config']['backend']`, and should be clarified when revising
the QJob class.