Revise anonymous register creation, tests
Revise the anonymous creation of `Registers`, simplifying it by using
`itertools.count` in the same spirit as the current implementation in
QuantumProgram. Fix a number of places where the arguments to the
constructor were passed in the wrong order, avoiding a ton of
deprecation warnings during the tests.

Reintroduce the ability of passing an empty circuit list to
QuantumProgram.compile(), which was lost during the transition.

Add new (and not complete) tests to `test_identifiers` that mimic the
tests for QuantumProgram anonymous features.