Unify handling of backend configurations
Revise `BaseBackend`, in order to unify the handling of the
`configuration` attribute. The default configuration for each local
backend is declared as a class attribute, and used by the base
initialization if it is not overridden during instantiation of the
subclasses. For the `IBMQBackend`, `configuration` is mandatory.
Drop the usage of ABC for simplicity, as the main benefits provided
by ABC are currently not being used.